requirements.txt
Django==2.1.4
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
psycopg2==2.7.6.1
PyJWT==1.7.1
pytz==2018.7
